+++ brt/src/jp/ne/sakura/uhideyuki/brt/brtsyn/LitInt.java
@@ -1,8 +1,8 @@
 package jp.ne.sakura.uhideyuki.brt.brtsyn;
 
 public class LitInt extends Literal { 
-    public Integer value; 
-    public LitInt(Integer x){ value = x; }
+    public long value; 
+    public LitInt(long x){ value = x; }
     public Boolean equals(Literal x){ return value == ((LitInt)x).value; }
 
     public String inspect(){
